Checkstyle: Fix package name violations
