+++ src/main/java/games/strategy/engine/framework/network/ui/ChangeGameToSaveGameClientAction.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.networkMaintenance;
+package games.strategy.engine.framework.network.ui;
 
 import java.awt.event.ActionEvent;
 import java.io.File;
