+++ src/main/java/games/strategy/engine/framework/network/ui/GetGameSaveClientAction.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.networkMaintenance;
+package games.strategy.engine.framework.network.ui;
 
 import java.awt.Component;
 import java.awt.Frame;
