+++ src/main/java/games/strategy/engine/framework/network/ui/SetPasswordAction.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.framework.networkMaintenance;
+package games.strategy.engine.framework.network.ui;
 
 import java.awt.BorderLayout;
 import java.awt.Component;
