+++ src/main/java/games/strategy/engine/framework/startup/mc/ClientModel.java
@@ -30,11 +30,11 @@ import games.strategy.engine.framework.GameObjectStreamFactory;
 import games.strategy.engine.framework.GameRunner;
 import games.strategy.engine.framework.IGameLoader;
 import games.strategy.engine.framework.message.PlayerListing;
-import games.strategy.engine.framework.networkMaintenance.ChangeGameOptionsClientAction;
-import games.strategy.engine.framework.networkMaintenance.ChangeGameToSaveGameClientAction;
-import games.strategy.engine.framework.networkMaintenance.ChangeToAutosaveClientAction;
-import games.strategy.engine.framework.networkMaintenance.GetGameSaveClientAction;
-import games.strategy.engine.framework.networkMaintenance.SetMapClientAction;
+import games.strategy.engine.framework.network.ui.ChangeGameOptionsClientAction;
+import games.strategy.engine.framework.network.ui.ChangeGameToSaveGameClientAction;
+import games.strategy.engine.framework.network.ui.ChangeToAutosaveClientAction;
+import games.strategy.engine.framework.network.ui.GetGameSaveClientAction;
+import games.strategy.engine.framework.network.ui.SetMapClientAction;
 import games.strategy.engine.framework.startup.launcher.IServerReady;
 import games.strategy.engine.framework.startup.login.ClientLogin;
 import games.strategy.engine.framework.startup.ui.ClientOptions;
