+++ src/main/java/games/strategy/engine/framework/startup/ui/ServerSetupPanel.java
@@ -30,10 +30,10 @@ import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.framework.IGame;
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
-import games.strategy.engine.framework.networkMaintenance.BanPlayerAction;
-import games.strategy.engine.framework.networkMaintenance.BootPlayerAction;
-import games.strategy.engine.framework.networkMaintenance.MutePlayerAction;
-import games.strategy.engine.framework.networkMaintenance.SetPasswordAction;
+import games.strategy.engine.framework.network.ui.BanPlayerAction;
+import games.strategy.engine.framework.network.ui.BootPlayerAction;
+import games.strategy.engine.framework.network.ui.MutePlayerAction;
+import games.strategy.engine.framework.network.ui.SetPasswordAction;
 import games.strategy.engine.framework.startup.launcher.ILauncher;
 import games.strategy.engine.framework.startup.launcher.ServerLauncher;
 import games.strategy.engine.framework.startup.login.ClientLoginValidator;
